fix(ci): resolve macOS build warnings and potential failures
This commit addresses several compiler warnings to stabilize the macOS CI build:
- Replaced deprecated juce::Font constructors with juce::FontOptions in FontManager.
- Fixed implicit float-to-int and sign conversion warnings in GridBackground, ThemeManager, and PluginProcessor.
- Suppressed unused parameter warnings in several UI classes using [[maybe_unused]].
- Removed an unused variable in WorkflowPage.

+++ src/PluginProcessor.h
@@ -77,7 +77,7 @@ public:
     //==============================================================================
     bool isNoteActive(int note) const
     {
-        return activeNotes[note].load(std::memory_order_relaxed);
+        return activeNotes[static_cast<size_t>(note)].load(std::memory_order_relaxed);
     }
 
     void triggerEnvelope(int index) { manualTriggerIndex = index; }
